// src/app/models/irmao.model.ts
export class Irmao{
     codigo: number;
     nome : string;
     email: string;
     telefone: string;
     sexo: string;
     indicador: boolean;
     microfonista: boolean;
     leitorSentinela: boolean;
     leitorEstudoLivro: boolean;
     sistemaSonoro: boolean;
     oracaoFinal: boolean;
     presidenteConferencia: boolean;
     carrinho: boolean;
     grupoId: number;
     congregacaoId: number;
     acessoAdmin: boolean;
     subirQuadro: boolean;
     atualizarAssistencia: boolean;
}